Restructuring the codebase to follow a more modular pattern Creating dedicated folders for each tool with separate schema and implementation files Cleaning up unnecessary comments while preserving important ones like shebang lines Following best practices for code commenting by avoiding obvious comments
