+++ src/tools/clear_cache/schema.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+
+// Schema for cache management
+export const schema = z.object({
+  endpoint: z.string().optional().describe("Optional specific endpoint to clear from cache. Leave empty to clear all."),
+});
