+++ src/tools/get_cache_stats/schema.ts
@@ -1,0 +1,4 @@
+import { z } from "zod";
+
+// Empty schema since this tool doesn't require any parameters
+export const schema = z.object({});
